# API/swagger.yaml
---
swagger: '2.0'
info:
  version: v1
  title: My API
  description: A simple example ASP.NET Core Web API
  termsOfService: None
  license:
    name: Use under LICX
    url: https://example.com/license
# host: petstore.swagger.io
# basePath: /v2
paths:
  "/api/games":
    get:
      tags:
      - Game
      operationId: ApiGamesGet
      consumes: []
      produces: []
      responses:
        '200':
          description: Success
    post:
      tags:
      - Game
      operationId: ApiGamesPost
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces: []
      parameters:
      - name: newGame
        in: body
        required: false
        schema:
          "$ref": "#/definitions/GameViewModel"
      responses:
        '200':
          description: Success
  "/api/games/{id}":
    get:
      tags:
      - Game
      operationId: ApiGamesByIdGet
      consumes: []
      produces: []
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        format: int32
      responses:
        '200':
          description: Success
  "/api/gameinfo":
    get:
      tags:
      - Game
      operationId: ApiGameinfoGet
      consumes: []
      produces: []
      responses:
        '200':
          description: Success
  "/api/players":
    get:
      tags:
      - Player
      operationId: ApiPlayersGet
      consumes: []
      produces: []
      responses:
        '200':
          description: Success
    post:
      tags:
      - Player
      operationId: ApiPlayersPost
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces: []
      parameters:
      - name: newPlayer
        in: body
        required: false
        schema:
          "$ref": "#/definitions/PlayerViewModel"
      responses:
        '200':
          description: Success
  "/api/players/{id}":
    get:
      tags:
      - Player
      operationId: ApiPlayersByIdGet
      consumes: []
      produces: []
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        format: int32
      responses:
        '200':
          description: Success
    put:
      tags:
      - Player
      operationId: ApiPlayersByIdPut
      consumes:
      - application/json-patch+json
      - application/json
      - text/json
      - application/*+json
      produces: []
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        format: int32
      - name: updatedPlayer
        in: body
        required: false
        schema:
          "$ref": "#/definitions/EditPlayerViewModel"
      responses:
        '200':
          description: Success
    delete:
      tags:
      - Player
      operationId: ApiPlayersByIdDelete
      consumes: []
      produces: []
      parameters:
      - name: id
        in: path
        required: true
        type: integer
        format: int32
      responses:
        '200':
          description: Success
  "/api/players/{pid1}/{pid2}":
    get:
      tags:
      - Player
      operationId: ApiPlayersByPid1ByPid2Get
      consumes: []
      produces: []
      parameters:
      - name: pid1
        in: path
        required: true
        type: integer
        format: int32
      - name: pid2
        in: path
        required: true
        type: integer
        format: int32
      responses:
        '200':
          description: Success
  "/api/players/bestteammate/{pid}":
    get:
      tags:
      - Player
      operationId: ApiPlayersBestteammateByPidGet
      consumes: []
      produces: []
      parameters:
      - name: pid
        in: path
        required: true
        type: integer
        format: int32
      responses:
        '200':
          description: Success
definitions:
  GameViewModel:
    type: object
    properties:
      id:
        format: int32
        type: integer
      date:
        type: string
      teamOneScore:
        format: int32
        type: integer
      teamTwoScore:
        format: int32
        type: integer
      teamOneList:
        type: array
        items:
          format: int32
          type: integer
      teamTwoList:
        type: array
        items:
          format: int32
          type: integer
  PlayerViewModel:
    type: object
    properties:
      name:
        type: string
  EditPlayerViewModel:
    type: object
    properties:
      name:
        type: string
      wins:
        format: int32
        type: integer
      losses:
        format: int32
        type: integer
      draws:
        format: int32
        type: integer
      gamesWon:
        format: int32
        type: integer
      gamesLost:
        format: int32
        type: integer
      attented:
        format: int32
        type: integer
host: virtserver.swaggerhub.com
basePath: /orriaxels/mballerss/1.0.0
schemes:
 - https
 - http